Removed .env and updated gitignore

+++ Controllers/review.js
@@ -1,24 +1,48 @@
 const Listing = require("../models/listing");
 const Review = require("../models/review");
 
-module.exports.createReview =async (req,res)=>{
-  let listing =await Listing.findById(req.params.id);
-  let newReview = new Review(req.body.review);
-newReview.author = req.user._id; 
-console.log(newReview);
-  listing.reviews.push(newReview);
+module.exports.createReview = async (req, res) => {
+    let { id } = req.params;
 
-  await newReview.save();
-  await listing.save();
-  req.flash("success","New Review Created!");
-res.redirect(`/listings/${lsiting_id}`);
+    let listing = await Listing.findById(id);
+
+    if (!listing) {
+        req.flash("error", "Listing not found!");
+        return res.redirect("/listings");
+    }
+
+    let newReview = new Review(req.body.review);
+
+    newReview.author = req.user._id;
+
+    listing.reviews.push(newReview);
+
+    await newReview.save();
+    await listing.save();
+
+    req.flash("success", "New Review Created!");
+
+    res.redirect(`/listings/${id}`);
 };
 
-module.exports.destroyReview= async (req,res)=>{
-    let{ id, reviewId }=req.params;
+module.exports.destroyReview = async (req, res) => {
+    let { id, reviewId } = req.params;
 
-    await Listing.findByIdAndUpdate(id, {$pull:{review: reviewId}})
+    let listing = await Listing.findById(id);
+
+    if (!listing) {
+        req.flash("error", "Listing not found!");
+        return res.redirect("/listings");
+    }
+
+    await Listing.findByIdAndUpdate(
+        id,
+        { $pull: { reviews: reviewId } }
+    );
+
     await Review.findByIdAndDelete(reviewId);
-  req.flash("success","New Review Deleted!");
+
+    req.flash("success", "Review Deleted!");
+
     res.redirect(`/listings/${id}`);
-  };
+};
